+++ packages/api/src/main/java/com/nsa/bt/scrumble/repositories/implementations/WorkspaceRepository.java
@@ -62,7 +62,7 @@ public class WorkspaceRepository implements IWorkspaceRepository {
             ps.setObject(4, getWorkspaceJsonbData(workspace));
             return ps;
         }, keyHolder);
-        workspace.setId(keyHolder.getKey().longValue());
+        workspace.setId(Math.toIntExact(keyHolder.getKey().longValue()));
         return workspace;
     }
